Moving ukvm:solo5_exec hypercall to a module
