+++ ukvm/ukvm.h
@@ -150,6 +150,7 @@ extern struct ukvm_module ukvm_module_core;
 extern struct ukvm_module ukvm_module_blk;
 extern struct ukvm_module ukvm_module_net;
 extern struct ukvm_module ukvm_module_gdb;
+extern struct ukvm_module ukvm_module_exec;
 
 /*
  * GDB specific functions to be implemented on all backends for all
